задание на 4 сделал

+++ js/main.js
@@ -1,3 +1,6 @@
+
+
+
 Vue.component('columns', {
     template: `
     <div class="list-notes">
@@ -9,13 +12,14 @@ Vue.component('columns', {
         <column-completed-tasks :cardList="cardsFour"></column-completed-tasks>
       </div>
     </div>
-            `,
+    `,
     data() {
         return {
             cardsOne: [],
             cardsTwo: [],
             cardsThree: [],
-            cardsFour: []
+            cardsFour: [],
+            clickedReason: null  // добавляем переменную для хранения выбранной причины возврата
         }
     },
     methods: {
@@ -27,8 +31,12 @@ Vue.component('columns', {
             this.deleteCard(card, this.cardsOne);
         },
         moveToThree(card) {
-            this.cardsThree.push(card);
-            this.deleteCard(card, this.cardsTwo);
+            if (card.reason.length > 0 && this.$root.clickedReason === null) {
+                alert('Необходимо выбрать причину возврата!');
+            } else {
+                this.cardsThree.push(card);
+                this.deleteCard(card, this.cardsTwo);
+            }
         },
         moveToFour(card) {
             card.completed = true;
@@ -46,6 +54,7 @@ Vue.component('columns', {
                 this.$set(card, 'reason', []);
             }
             card.reason.push(reason);
+            card.checked = false;
         },
         updateDeadlines(deadline) {
             this.cardsOne.forEach(card => card.deadline = deadline);
@@ -58,9 +67,10 @@ Vue.component('columns', {
                 list.splice(index, 1);
             }
         },
-
     }
 });
+
+
 
 Vue.component('create-card', {
     template: `
@@ -156,22 +166,28 @@ Vue.component('column-tasks-work', {
         cardList: Array,
     },
     template: `
-            <div class="col">
-                <card-form class="column"
-                    v-for="card in cardList"
-                    :key="card.title"
-                    :card="card"
-                    @move-card="moveCard"
-                    :column="2"> <!-- Pass the column number -->
-                </card-form>
-            </div>
-            `,
+        <div class="col">
+            <card-form class="column"
+                v-for="card in cardList"
+                :key="card.title"
+                :card="card"
+                @move-card="moveCard"
+                :column="2"> <!-- Pass the column number -->
+            </card-form>
+        </div>
+    `,
     methods: {
         moveCard(card) {
-            this.$emit('move-to-three', card);
+            if (card.reason.length > 0 && !card.checked) {
+                alert('Необходимо выбрать причину возврата и отметить чекбокс!');
+            } else {
+                this.$emit('move-to-three', card);
+            }
         }
     }
 })
+
+
 
 
 Vue.component('column-testing', {
@@ -217,7 +233,7 @@ Vue.component('column-completed-tasks', {
                     v-for="card in cardList"
                     :key="card.title"
                     :card="card"
-                    :column="4"> <!-- Pass the column number -->
+                    :column="4"> 
                 </card-form>
             </div>
             `,
@@ -266,6 +282,7 @@ Vue.component('card-edit', {
         }
     }
 });
+
 Vue.component('card-form', {
     template: `
     <div class="cardOne">
@@ -275,11 +292,12 @@ Vue.component('card-form', {
         <p>Дедлайн:{{ card.deadline }}</p>
         <p v-if="card.dateEdit != null">Редактирование: {{ card.dateEdit }}</p>
         <p v-if="last != true && card.reason.length  > 0" >
-          <span v-for="reas in card.reason" :class="{ 'strike-through': clickedReason === reas }" @click="toggleReason(reas)">{{ reas }}</span>
+        <span v-for="reas in card.reason" :class="{ 'strike-through': clickedReason === reas }" @click="column === 2 && toggleReason(reas)">
+            {{ reas }}
+        <input type="checkbox" v-model="card.checked" :disabled="column !== 2">
+        </span>
+
         </p>
-<!--        <ul>-->
-<!--          <li v-for="reas in card.reason" :class="{ 'strike-through': clickedReason === reas }" @click="toggleReason(reas)">{{ reas }}</li>-->
-<!--        </ul>-->
         <p>Дата создания:{{card.dateCreate}}</p>
         <p v-if="card.completed != null">Карточка:  {{ card.completed ? 'Просрочен' : 'Выполнен' }}</p>
         <button class="peremestit" type="button" @click="moveCard" v-if="card.completed === null">
@@ -330,6 +348,7 @@ Vue.component('card-form', {
         },
         returnCard() {
             if (this.reason) {
+                this.card.reason = [];
                 this.$emit('return-card', this.card, this.reason);
                 this.reason = '';
             }
@@ -403,3 +422,11 @@ let app = new Vue({
         }
     }
 });
+
+
+
+
+
+
+
+
